+++ packages/stylist-core/src/ast/selector.rs
@@ -1,5 +1,7 @@
 use std::borrow::Cow;
 
+use serde::{Deserialize, Serialize};
+
 use super::{StringFragment, StyleContext, ToStyleStr};
 
 /// A CSS Selector.
@@ -8,7 +10,7 @@ use super::{StringFragment, StyleContext, ToStyleStr};
 /// ```css
 /// div[attr="val"].my-class#some-id
 /// ```
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
+#[derive(Debug, Clone, PartialEq, Eq, Hash, Serialize, Deserialize)]
 pub struct Selector {
     pub fragments: Cow<'static, [StringFragment]>,
 }
@@ -22,7 +24,7 @@ impl ToStyleStr for Selector {
         }
 
         if let Some(ref m) = ctx.class_name {
-            let scoped_class = format!(".{}", m);
+            let scoped_class = format!(".{m}");
             // If contains current selector or root pseudo class, replace them with class name.
             if joined_s.contains('&') || joined_s.contains(":root") {
                 w.push_str(