+++ packages/stylist/src/yew/provider.rs
@@ -6,6 +6,7 @@ use crate::manager::StyleManager;
 #[derive(Properties, Clone, Debug, PartialEq)]
 pub struct ManagerProviderProps {
     pub manager: StyleManager,
+    #[prop_or_default]
     pub children: Children,
 }
 
@@ -36,10 +37,42 @@ pub struct ManagerProviderProps {
 /// }
 /// ```
 #[function_component(ManagerProvider)]
-pub fn manager_provider(props: &ManagerProviderProps) -> Html {
-    html! {
-        <ContextProvider<StyleManager> context={props.manager.clone()}>
-            {props.children.clone()}
-        </ContextProvider<StyleManager>>
+pub fn manager_provider(props: &ManagerProviderProps) -> HtmlResult {
+    let ManagerProviderProps { manager, children } = props.clone();
+
+    #[cfg(any(feature = "ssr", feature = "hydration"))]
+    {
+        use crate::manager::StyleData;
+
+        let _manager = manager.clone();
+        let _style_data = use_transitive_state!(
+            move |_| -> StyleData {
+                _manager
+                    .style_data()
+                    .map(|m| m.clone())
+                    .unwrap_or_else(StyleData::new)
+            },
+            ()
+        )?;
+
+        #[cfg(feature = "hydration")]
+        {
+            // We must load the styles immediately before child components are rendered.
+            let manager = manager.clone();
+            use_memo(
+                move |manager| {
+                    if let Some(m) = _style_data {
+                        manager.load_style_data(m.as_ref());
+                    }
+                },
+                manager,
+            );
+        }
     }
+
+    Ok(html! {
+        <ContextProvider<StyleManager> context={manager}>
+            {children}
+        </ContextProvider<StyleManager>>
+    })
 }